requirements.txt
fastapi
uvicorn
pydub
gTTS
python-dotenv
websockets
jinja2
aiofiles
requests
translate